move set_interp_out_fps from args to frameinterp
